Binding custom shaders works for super simple shaders, some standard attributes and uniforms are automatically bound if present in the shader.
ofFBOs with depth are not working.

gles2-bc is now in the libs folder, but could be moved back out to the ofsGLES2Renderer addon folder again as we are only using the file "OpenGLESDefines.h".

ofConstants.h currently includes OpenGLES/ES1/gl.h even in OPENGLES_VERSION_2 to make it compile, there is a bunch of stuff that needs to be redirected to the renderer or #ifdef'd out in ES2 mode, will add to the todo-list.